backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.26.0
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Image Processing
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
